requirements.txt
streamlit
langchain
langchain-community
openai
requests
python-dotenv
wikipedia